test(gateway): split node pairing memo coverage
